+++ src/Mesh_t.cu
@@ -20,13 +20,13 @@ Mesh_t::Mesh_t(size_t N_elements, int initial_N, deviceFloat x_min, deviceFloat 
         elements_numBlocks_((N_elements_ + elements_blockSize - 1) / elements_blockSize),
         faces_numBlocks_((N_faces_ + faces_blockSize - 1) / faces_blockSize),
         host_delta_t_array_(new deviceFloat[elements_numBlocks_]),
-        host_refine_array_(new unsigned int[elements_numBlocks_]) {
+        host_refine_array_(new unsigned long[elements_numBlocks_]) {
 
     // CHECK N_faces = N_elements only for periodic BC.
     cudaMalloc(&elements_, N_elements_ * sizeof(Element_t));
     cudaMalloc(&faces_, N_faces_ * sizeof(Face_t));
     cudaMalloc(&device_delta_t_array_, elements_numBlocks_ * sizeof(deviceFloat));
-    cudaMalloc(&device_refine_array_, elements_numBlocks_ * sizeof(unsigned int));
+    cudaMalloc(&device_refine_array_, elements_numBlocks_ * sizeof(unsigned long));
 
     SEM::build_elements<<<elements_numBlocks_, elements_blockSize>>>(N_elements_, initial_N_, elements_, x_min, x_max);
     SEM::build_faces<<<faces_numBlocks_, faces_blockSize>>>(N_faces_, faces_); // CHECK
@@ -346,7 +346,7 @@ deviceFloat Mesh_t::get_delta_t(const deviceFloat CFL) {
 
 void Mesh_t::adapt(const deviceFloat* nodes, const deviceFloat* barycentric_weights) {
     SEM::reduce_refine<elements_blockSize><<<elements_numBlocks_, elements_blockSize>>>(N_elements_, elements_, device_refine_array_);
-    cudaMemcpy(host_refine_array_, device_refine_array_, elements_numBlocks_ * sizeof(unsigned int), cudaMemcpyDeviceToHost);
+    cudaMemcpy(host_refine_array_, device_refine_array_, elements_numBlocks_ * sizeof(unsigned long), cudaMemcpyDeviceToHost);
 
     unsigned int additional_elements = 0;
     for (int i = 0; i < elements_numBlocks_; ++i) {
@@ -358,7 +358,7 @@ void Mesh_t::adapt(const deviceFloat* nodes, const deviceFloat* barycentric_weig
         return;
     }
 
-    cudaMemcpy(device_refine_array_, host_refine_array_, elements_numBlocks_ * sizeof(unsigned int), cudaMemcpyHostToDevice);
+    cudaMemcpy(device_refine_array_, host_refine_array_, elements_numBlocks_ * sizeof(unsigned long), cudaMemcpyHostToDevice);
 
     Element_t* new_elements;
     Face_t* new_faces;
@@ -384,12 +384,12 @@ void Mesh_t::adapt(const deviceFloat* nodes, const deviceFloat* barycentric_weig
     delete[] host_delta_t_array_;
     delete[] host_refine_array_;
     host_delta_t_array_ = new deviceFloat[elements_numBlocks_];
-    host_refine_array_ = new unsigned int[elements_numBlocks_]; 
+    host_refine_array_ = new unsigned long[elements_numBlocks_]; 
 
     cudaFree(device_delta_t_array_);
     cudaFree(device_refine_array_);
     cudaMalloc(&device_delta_t_array_, elements_numBlocks_ * sizeof(deviceFloat));
-    cudaMalloc(&device_refine_array_, elements_numBlocks_ * sizeof(unsigned int));
+    cudaMalloc(&device_refine_array_, elements_numBlocks_ * sizeof(unsigned long));
 
     print();
 }
